requirements.txt
pyyaml==5.3
selenium==3.141.0
pandas==1.0.1
matplotlib==3.1.3
seaborn==0.10.0
scikit-learn==0.22.2.post1
imbalanced-learn==0.6.2